largest-contentful-paint/progressively-loaded-image.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: progressively loaded image</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/largest-contentful-paint-helpers.js"></script>
<script>
  async_test(function (t) {
    assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");
    const beforeLoad = performance.now();
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const imgElement = document.getElementById('image_id');
        const imageSrc = window.location.origin + imgElement.getAttribute('src');
        const imageWidth = imgElement.naturalWidth;
        const imageHeight = imgElement.naturalHeight;
        const size = imageWidth * imageHeight;
        const lcp = entry.loadTime;
        calculateResourceLoadTime(imageSrc);
        assert_greater_than(lcp, resourceLoadTime,'LCP event fired');
        t.step_timeout(() => {
          checkImage(entry, imageSrc, 'image_id', size, beforeLoad);
          t.done();
        }, 1000);
      })
    );
    observer.observe({type: 'largest-contentful-paint', buffered: true});
    function calculateResourceLoadTime(imageSrc) {
    const resourceEntry = performance.getEntriesByName(imageSrc)[0];
    if (resourceEntry) {
      resourceLoadTime = resourceEntry.responseEnd - resourceEntry.startTime;
    } else {
      assert_equals(resourceEntry, 'Resource not found');
    }
  }
}, 'Same-origin image is observable.');
</script>
<img src='/images/arrow-oriented-upright.jpg?pipe=trickle(17280:d1)' id='image_id'/>
</body>
</html>
